Created common project and deleted floating classes from src/shared
